Extracted facet processing logic of CMDIParserVTDXML into FacetProcessor class
